# docs/swagger.yaml
basePath: /
definitions:
  controllers.OrderRequest:
    properties:
      order_items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      table_id:
        type: string
    required:
    - order_items
    - table_id
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.Note:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      note_id:
        type: string
      text:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  models.Order:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      order_date:
        type: string
      order_id:
        type: string
      order_items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      table_id:
        type: string
      updatedAt:
        type: string
    required:
    - order_date
    type: object
  models.OrderItem:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      food_id:
        type: string
      id:
        type: integer
      order_id:
        type: string
      order_item_id:
        type: string
      quantity:
        type: string
      unit_price:
        type: number
      updatedAt:
        type: string
    required:
    - food_id
    - order_id
    - quantity
    - unit_price
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: |-
    This is a comprehensive restaurant management system API built with Go and Gin framework
    Features include user management, food menus, table reservations, orders, and invoicing
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Restaurant Management API
  version: "1.0"
paths:
...
  /notes:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all notes
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all notes
      tags:
      - Notes
    post:
      consumes:
      - application/json
      description: Create a new note with the provided information
      parameters:
      - description: Note object
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/models.Note'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new note
      tags:
      - Notes
  /notes/{note_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific note by note_id
      parameters:
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Note'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get note by ID
      tags:
      - Notes
    put:
      consumes:
      - application/json
      description: Update an existing note by note_id
      parameters:
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: string
      - description: Note object
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/models.Note'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a note
      tags:
      - Notes
  /orderitems:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all order items
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all order items
      tags:
      - OrderItems
    post:
      consumes:
      - application/json
      description: Add a new item to an existing order
      parameters:
      - description: Order item object
        in: body
        name: order_item
        required: true
        schema:
          $ref: '#/definitions/models.OrderItem'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new order item
      tags:
      - Orders
...
  /orders:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all orders with their order items
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all orders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Create a new order with order items. Automatically validates food
        items and calculates prices.
      parameters:
      - description: Order with items
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/controllers.OrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - Orders
  /orders/{order_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific order by order_id with its order items
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - Orders
    put:
      consumes:
      - application/json
      description: Update an existing order by order_id. Cannot update if the order
        is already paid.
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      - description: Order object
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.Order'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update an order
      tags:
      - Orders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...