# models/orderItemModel.go
package models

import "gorm.io/gorm"

type OrderItem struct {
	gorm.Model
	Quantity      *string  `json:"quantity" validate:"required,eq=S|eq=M|eq=L"`
	Unit_price    *float64 `json:"unit_price" validate:"required"`
	Food_id       *string  `json:"food_id" validate:"required"`
	Order_item_id string   `json:"order_item_id"`
	Order_id      string   `json:"order_id" validate:"required"`
}